kv/bulk: extract flush implementation into sst_adder.go
This commit moves parts of the SSTBatcher that will become the sstAdder
into a stand alone file.

Release note: none
Part of: #146828
